:truck: Refactored and moved code
