+++ scss/base/_resets.scss
@@ -32,11 +32,6 @@ tr:hover {
 th {
   background-color: $main-color;
   color: #ffffff;
-  text-align: center !important;
-}
-
-td {
-  text-align: center !important;
 }
 
 .table th {
